Changed the author in the model to be foreign key.....
So when the user creates a recipe it'll save with the current user's username. However I had to delete the last database because the author in the schema was a varchar, and couldn't recognize the foreign key. Deleted db.sqlite3 file and all the migrations except __init__ file and made migrations again.